# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VBUY — Premium Ghanaian E-Commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/customers:
    get:
      operationId: listAdminCustomers
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomerListResponse"
  /admin/customers/{userId}:
    get:
      operationId: getAdminCustomer
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Customer profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminCustomerDetail"
    patch:
      operationId: updateAdminCustomer
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [active, flagged, disabled] }
                email: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
                required: [userId, status]
                properties:
                  userId: { type: string }
                  email: { type: ["string", "null"] }
                  status: { type: string }
  /admin/customers/{userId}/notes:
    post:
      operationId: addAdminCustomerNote
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: userId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string }
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminNote"
...
  /admin/content:
    get:
      operationId: listAdminContent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Content blocks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContentBlock"
    post:
      operationId: createAdminContent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentBlockInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentBlock"
  /admin/content/{id}:
    patch:
      operationId: updateAdminContent
      tags: [admin]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContentBlockInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContentBlock"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Address:
      type: object
      properties:
        id:             { type: integer }
        digitalAddress: { type: string }
        region:         { type: string }
        district:       { type: string }
        notes:          { type: ["string", "null"] }
        userId:         { type: ["string", "null"] }
      required: [id, digitalAddress, region, district]
...
    AdminOrderSummary:
      type: object
      properties:
        id: { type: integer }
        status: { type: string }
        paymentMethod: { type: string }
        paymentStatus: { type: string }
        total: { type: number }
        userId: { type: ["string", "null"] }
        carrier: { type: ["string", "null"] }
        trackingNumber: { type: ["string", "null"] }
        momoReference: { type: ["string", "null"] }
        paymentSubmittedAt: { type: ["string", "null"] }
        duplicateMomoReference: { type: boolean }
        createdAt: { type: string }
        region: { type: ["string", "null"] }
        district: { type: ["string", "null"] }
      required: [id, status, paymentMethod, paymentStatus, total, createdAt, duplicateMomoReference]
...
    AdminNote:
      type: object
      properties:
        id: { type: integer }
        authorUserId: { type: string }
        body: { type: string }
        createdAt: { type: string }
      required: [id, authorUserId, body, createdAt]
...
    AdminCustomerSummary:
      type: object
      properties:
        userId: { type: string }
        email: { type: ["string", "null"] }
        status: { type: string }
        orderCount: { type: integer }
        totalSpent: { type: number }
        lastOrderAt: { type: ["string", "null"] }
      required: [userId, status, orderCount, totalSpent]
    AdminCustomerListResponse:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/AdminCustomerSummary"
        total: { type: integer }
        page: { type: integer }
        totalPages: { type: integer }
      required: [customers, total, page, totalPages]
    AdminCustomerDetail:
      type: object
      properties:
        userId: { type: string }
        email: { type: ["string", "null"] }
        status: { type: string }
        orderCount: { type: integer }
        totalSpent: { type: number }
        orders:
          type: array
          items:
            $ref: "#/components/schemas/AdminOrderSummary"
        addresses:
          type: array
          items:
            $ref: "#/components/schemas/Address"
        notes:
          type: array
          items:
            $ref: "#/components/schemas/AdminNote"
      required: [userId, status, orderCount, totalSpent, orders, addresses, notes]
...
    ContentBlock:
      type: object
      properties:
        id: { type: integer }
        key: { type: string }
        title: { type: string }
        body: { type: string }
        imageUrl: { type: ["string", "null"] }
        ctaLabel: { type: ["string", "null"] }
        ctaHref: { type: ["string", "null"] }
        sortOrder: { type: integer }
        published: { type: boolean }
        updatedAt: { type: string }
      required: [id, key, title, body, sortOrder, published, updatedAt]
    ContentBlockInput:
      type: object
      properties:
        key: { type: string }
        title: { type: string }
        body: { type: string }
        imageUrl: { type: ["string", "null"] }
        ctaLabel: { type: ["string", "null"] }
        ctaHref: { type: ["string", "null"] }
        sortOrder: { type: integer }
        published: { type: boolean }